Fix OLS detection
The operator is currently looking for the OLS CSV only on the namespace
where the operator is installed, but its not going to find it because
OLS is usually installed in the `openshift-lightspeed` namespace.

In this patch we search ALL namespaces for the CSV, after all OLS is a
cluster wide operator, so it doesn't matter where we find it.