Add OLS operator installation
Implement automatic installation, version checking ,and lifecycle
management of the OpenShift Lightspeed operator. Add support for both
user-installed and instance-owned installation modes with proper
ownership tracking.

If the OLS operator has already been installed then the OpenStack
Lightspeed operator checks it version. If the version is valied
OpenStack Lightspeed utilizes the preinstalled OLS operator.

If the OLS operator has not been already installed then the operator
installs it. In this mode when the OpenStackLightspeed instance is
created the OLS operator is automatically uninstalled as well. Whereas
in the preinstalled OLS operator mode the OLS oeprator stays even after
the removal of the OpenStackLightspeed instance.