Add OLSConfig ping mechanism
Introduce a ping function that adds a random label to OLSConfig
resources to trigger reconciliation by the OpenShift Lightspeed
operator. This ensures the operator updates the Status field when
the OLSConfig is not ready.

This addresses a current limitation where the OLS operator does
not automatically update the OLSConfig status when installed in
the openstack-lightspeed namespace.