Leverage the autodiscovery indexID feature
This patch leverage the autodiscovery indexID feature from OLS.

Prior to this patch, our code started a temporary container to extract
the indexID value from an env variable, this is no longer needed for OLS
and the code can be removed.

Now, OLS levarages the rag of the RAG images as the indexID for the
vector database. Since our images already matches the tag with the
IndexID that was used during build time, we no longer need to set it in
OLSConfig.

Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>